Move size_of() to the separate file
Because this is more about IO rather than type system itself
